requirements.txt
openai==1.55.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
ollama==0.4.7
pandas
openpyxl
requests==2.32.3
antlr4-python3-runtime==4.13.2
jinja2
